refactor: split engineer_loop.rs into sub-modules
Split the 2346-line monolithic engineer_loop.rs into a directory module
with focused sub-modules, each ≤400 impl lines:

- types.rs (295 lines): structs, enums, constants, analyze_objective,
  parse helpers
- selection.rs (332 lines): select_engineer_action
- execution.rs (359 lines): execute_engineer_action, run_command,
  CommandOutput, helpers
- verification.rs (433 lines): verify_engineer_action and verification
  helpers
- review_persist.rs (257 lines): run_optional_review,
  compute_diff_for_review, persist_engineer_loop_artifacts
- mod.rs (365 impl + 381 test lines): core loop, workspace inspection,
  re-exports, all 29 existing tests

All public API paths (crate::engineer_loop::X) preserved via re-exports.
All 656 lib tests pass. Clippy clean.
